+++ Cargo.toml
@@ -10,7 +10,8 @@ guard = "0.5.1"
 bimap = "0.6.2"
 config = "0.13.1"
 voskrust = { git = "https://github.com/tikitko/voskrust.git" }
-songbird = "0.2.2"
+songbird = "0.3.0"
 fon = "0.6.0"
-serenity = { version = "0.10", default-features = false, features = ["builder", "cache", "client", "gateway", "model", "utils", "rustls_backend", "collector", "voice"] }
+serenity = { version = "0.11", default-features = false, features = ["builder", "cache", "client", "gateway", "model", "utils", "rustls_backend", "collector", "voice"] }
 tokio = { version = "1.0", features = ["macros", "rt-multi-thread"] }
+ngrammatic = "0.4.0"
